[Incubator][VC]Clean up - Refactor resource controller - Part 3
