+++ incubator/virtualcluster/pkg/syncer/resources/serviceaccount/controller.go
@@ -18,22 +18,23 @@ package serviceaccount
 
 import (
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	coreinformers "k8s.io/client-go/informers/core/v1"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
+	listersv1 "k8s.io/client-go/listers/core/v1"
+	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog"
 
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/constants"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/conversion"
 	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/manager"
 	mc "github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/mccontroller"
-	"github.com/kubernetes-sigs/multi-tenancy/incubator/virtualcluster/pkg/syncer/reconciler"
 )
 
 type controller struct {
-	client                               v1core.CoreV1Interface
-	saInformer                           coreinformers.ServiceAccountInformer
+	// super master sa client
+	saClient v1core.CoreV1Interface
+	// super master sa lister/synced function
+	saLister listersv1.ServiceAccountLister
+	saSynced cache.InformerSynced
+	// Connect to all tenant master sa informers
 	multiClusterServiceAccountController *mc.MultiClusterController
 }
 
@@ -43,7 +44,7 @@ func Register(
 	controllerManager *manager.ControllerManager,
 ) {
 	c := &controller{
-		client: client,
+		saClient: client,
 	}
 
 	// Create the multi cluster secret controller
@@ -54,6 +55,8 @@ func Register(
 		return
 	}
 	c.multiClusterServiceAccountController = multiClusterSecretController
+	c.saLister = saInformer.Lister()
+	c.saSynced = saInformer.Informer().HasSynced
 
 	controllerManager.AddController(c)
 }
@@ -62,90 +65,7 @@ func (c *controller) StartUWS(stopCh <-chan struct{}) error {
 	return nil
 }
 
-func (c *controller) StartDWS(stopCh <-chan struct{}) error {
-	return c.multiClusterServiceAccountController.Start(stopCh)
-}
-
 func (c *controller) StartPeriodChecker(stopCh <-chan struct{}) {
-}
-
-func (c *controller) Reconcile(request reconciler.Request) (reconciler.Result, error) {
-	klog.Infof("reconcile service account %s/%s %s event for cluster %s", request.Namespace, request.Name, request.Event, request.Cluster.Name)
-
-	switch request.Event {
-	case reconciler.AddEvent:
-		err := c.reconcileServiceAccountCreate(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.ServiceAccount))
-		if err != nil {
-			klog.Errorf("failed reconcile service account %s/%s CREATE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	case reconciler.UpdateEvent:
-		err := c.reconcileServiceAccountUpdate(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.ServiceAccount))
-		if err != nil {
-			klog.Errorf("failed reconcile service account %s/%s CREATE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	case reconciler.DeleteEvent:
-		err := c.reconcileServiceAccountRemove(request.Cluster.Name, request.Namespace, request.Name, request.Obj.(*v1.ServiceAccount))
-		if err != nil {
-			klog.Errorf("failed reconcile service account %s/%s DELETE of cluster %s %v", request.Namespace, request.Name, request.Cluster.Name, err)
-			return reconciler.Result{Requeue: true}, err
-		}
-	}
-	return reconciler.Result{}, nil
-}
-
-func (c *controller) reconcileServiceAccountCreate(cluster, namespace, name string, secret *v1.ServiceAccount) error {
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, namespace)
-	// just mark the default service account as a tenant related resource.
-	// service account controller will create it for us.
-	if name == "default" {
-		sa, err := c.client.ServiceAccounts(targetNamespace).Get("default", metav1.GetOptions{})
-		if err != nil {
-			// maybe the sa is not created, retry
-			return err
-		}
-
-		if len(sa.Annotations) == 0 {
-			sa.Annotations = make(map[string]string)
-		}
-		sa.Annotations[constants.LabelCluster] = cluster
-		_, err = c.client.ServiceAccounts(targetNamespace).Update(sa)
-		return err
-	}
-	newObj, err := conversion.BuildMetadata(cluster, targetNamespace, secret)
-	if err != nil {
-		return err
-	}
-
-	pServiceAccount := newObj.(*v1.ServiceAccount)
-	// set to empty and token controller will regenerate one.
-	pServiceAccount.Secrets = nil
-
-	_, err = c.client.ServiceAccounts(targetNamespace).Create(pServiceAccount)
-	if errors.IsAlreadyExists(err) {
-		klog.Infof("service account %s/%s of cluster %s already exist in super master", namespace, name, cluster)
-		return nil
-	}
-	return err
-}
-
-func (c *controller) reconcileServiceAccountUpdate(cluster, namespace, name string, secret *v1.ServiceAccount) error {
-	// do nothing.
-	return nil
-}
-
-func (c *controller) reconcileServiceAccountRemove(cluster, namespace, name string, secret *v1.ServiceAccount) error {
-	targetNamespace := conversion.ToSuperMasterNamespace(cluster, namespace)
-	opts := &metav1.DeleteOptions{
-		PropagationPolicy: &constants.DefaultDeletionPolicy,
-	}
-	err := c.client.ServiceAccounts(targetNamespace).Delete(name, opts)
-	if errors.IsNotFound(err) {
-		klog.Warningf("service account %s/%s of cluster %s not found in super master", namespace, name, cluster)
-		return nil
-	}
-	return err
 }
 
 func (c *controller) AddCluster(cluster mc.ClusterInterface) {
